+++ CHANGELOG.md
@@ -6,6 +6,20 @@ This project adheres to [Semantic Versioning](https://semver.org), except that â
 
 Do not manually edit this file. It will be automatically updated when a new release is published.
 
+## 0.2.0
+_26 May 2022_
+
+* Fix dependency reference from c2patool crate to c2pa crate ([#21](https://github.com/contentauth/c2pa-rs/pull/21))
+* (MINOR) Detailed API review for Ingredient struct ([#17](https://github.com/contentauth/c2pa-rs/pull/17))
+
+
+## 0.1.3
+_26 May 2022_
+
+* Publish c2patool crate ([#20](https://github.com/contentauth/c2pa-rs/pull/20))
+* Improve documentation ([#14](https://github.com/contentauth/c2pa-rs/pull/14))
+
+
 ## 0.1.2
 _26 May 2022_